Refactor : id값으로 존재 여부 확인 시 findById대신 existsById로 확인하는 것으로 수정

+++ src/main/java/com/miracle/userservice/service/InterviewServiceImpl.java
@@ -22,7 +22,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Transactional
 @Service
-public class InterviewServiceImpl implements InterviewService{
+public class InterviewServiceImpl implements InterviewService {
 
     private final InterviewRepository interviewRepository;
     private final UserRepository userRepository;
@@ -31,7 +31,7 @@ public class InterviewServiceImpl implements InterviewService{
     @Transactional(readOnly = true)
     @Override
     public InterviewResponseDto getInterviews(Long interviewId) {
-        if(interviewId == null) return null;
+        if (interviewId == null) return null;
 
         Optional<Interview> interviewOpt = interviewRepository.findById(interviewId);
         Interview interview = interviewOpt.orElseThrow(() -> new NoSuchInterviewException("400_1", "면접 정보가 존재하지 않습니다."));
@@ -44,25 +44,24 @@ public class InterviewServiceImpl implements InterviewService{
         Objects.requireNonNull(userId, "User id is null");
         Objects.requireNonNull(dto, "InterviewPostRequestDto is null");
 
-        User user = userRepository.findById(userId).orElseThrow();
-
-        ApplicationLetter applicationLetter = applicationLetterRepository.findById(dto.getApplicationLetterId()).orElseThrow(() -> new NoSuchApplicationLetterException("400_4", "지원서가 존재하지 않습니다."));
-        Optional<Interview> interviewOpt = interviewRepository.findByApplicationLetterId(dto.getApplicationLetterId());
-
-        if(interviewOpt.isPresent()) {
-            throw new DuplicateInterviewException("400_5", "면접 정보가 이미 존재합니다.");
+        boolean result = interviewRepository.existsByApplicationLetterId(dto.getApplicationLetterId());
+        if (result) {
+            throw new DuplicateInterviewException("400_3", "면접 정보가 이미 존재합니다.");
         }
+
+        if (dto.getQnaList().size() > 5) {
+            throw new OverflowException("406", "면접 정보는 최대 5개까지 저장 가능합니다.");
+        }
+
+        User user = userRepository.findById(userId).orElseThrow();
+        ApplicationLetter applicationLetter = applicationLetterRepository.findById(dto.getApplicationLetterId()).orElseThrow(() -> new NoSuchApplicationLetterException("400_4", "지원서가 존재하지 않습니다."));
 
         Interview interview = Interview.builder()
                 .user(user)
                 .applicationLetter(applicationLetter)
                 .build();
 
-        dto.getQnaList().forEach(interview::addQna);
-
-        if(dto.getQnaList().size() > 5) {
-            throw new OverflowException("406", "면접 정보는 최대 5개까지 저장 가능합니다.");
-        }
+        interview.getQnaList().addAll(dto.getQnaList());
 
         interviewRepository.save(interview);
 
@@ -74,17 +73,20 @@ public class InterviewServiceImpl implements InterviewService{
         Objects.requireNonNull(interviewId, "interview id is null");
         Objects.requireNonNull(dto, "InterviewPostRequestDto is null");
 
-        applicationLetterRepository.findById(dto.getApplicationLetterId()).orElseThrow(() -> new NoSuchApplicationLetterException("400_4", "지원서가 존재하지 않습니다."));
-
-        Optional<Interview> interviewOpt = interviewRepository.findById(interviewId);
-        Interview interview = interviewOpt.orElseThrow(() -> new NoSuchInterviewException("400_5", "면접 정보가 존재하지 않습니다."));
-
-        interview.getQnaList().clear();
-        dto.getQnaList().forEach(interview::addQna);
-
-        if(dto.getQnaList().size() > 5) {
+        if (dto.getQnaList().size() > 5) {
             throw new OverflowException("406", "면접 정보는 최대 5개까지 저장 가능합니다.");
         }
+
+        boolean result = applicationLetterRepository.existsById(dto.getApplicationLetterId());
+        if (!result) {
+            throw new NoSuchApplicationLetterException("400_3", "지원서가 존재하지 않습니다.");
+        }
+
+        Optional<Interview> interviewOpt = interviewRepository.findById(interviewId);
+        Interview interview = interviewOpt.orElseThrow(() -> new NoSuchInterviewException("400_4", "면접 정보가 존재하지 않습니다."));
+
+        interview.getQnaList().clear();
+        interview.getQnaList().addAll(dto.getQnaList());
 
         return true;
     }
@@ -93,7 +95,11 @@ public class InterviewServiceImpl implements InterviewService{
     public boolean deleteInterview(Long interviewId) {
         Objects.requireNonNull(interviewId, "Interview id is null");
 
-        interviewRepository.findById(interviewId).orElseThrow(() -> new NoSuchInterviewException("400_1", "면접 정보가 존재하지 않습니다."));
+        boolean result = interviewRepository.existsById(interviewId);
+        if (!result) {
+            throw new NoSuchInterviewException("400_1", "면접 정보가 존재하지 않습니다.");
+        }
+
         interviewRepository.deleteById(interviewId);
 
         return true;
